aws_backend/lambda/requirements.txt
boto3
requests
google-auth
